[tests] consolidate duplicate user fixtures in frontend tests
